+++ test/config_generator_test.rb
@@ -20,9 +20,10 @@ class ConfigGeneratorTest < Minitest::Test
       end
     end
 
+    end_line = 405
     assert_yaml "test_generate_with_tools.yml",
                 tools: tools,
-                comment_out_lines: [9..404, 407..411, 414..418]
+                comment_out_lines: [9..end_line, (end_line + 3)..(end_line + 7), (end_line + 10)..(end_line + 14)]
   end
 
   private